optimize the components to reduce re-render
To-do form and to-do items are optimized by using memoization to reduce re-renders as memoized component does not re-render when the parent component re-renders.

All handlers are also cached by using useCallBack hook, which allows the function to be created once and used across renders. Else, a new handler will be created on every render.